+++ src/hooks.ts
@@ -1,5 +1,6 @@
 import {
     ActorPF2e,
+    ActorSheetPF2e,
     CharacterPF2e,
     CreaturePF2e,
     CreatureSheetPF2e,
@@ -16,6 +17,7 @@ import { createRefinedItemDialog } from "./app/refined-item-create";
 import { MODULE_ID } from "./module";
 import { AfterCombatDialog } from "./app/after-combat-dialog";
 import { AutomaticRefinementProgression } from "@src/modules/automatic-refinement-progression";
+import { DynamicStyles } from "@src/modules/dynamic-styles";
 
 export class ModuleHooks {
     static registerAllHandlers() {
@@ -41,6 +43,8 @@ export class ModuleHooks {
         if (AutomaticRefinementProgression.isEnabled) {
             this.updateItemsOnLevelChangeARP();
         }
+
+        this.highlightItemsWithMatchingMaterials();
     }
 
     static addMonsterPartButton() {
@@ -232,6 +236,44 @@ export class ModuleHooks {
                 MonsterPart.fromCreature(actor);
             },
         );
+    }
+
+    static highlightItemsWithMatchingMaterials() {
+        Hooks.on("renderActorSheetPF2e", (sheet: ActorSheetPF2e<ActorPF2e>) => {
+            const actor = sheet.document;
+            if (!actor.isOfType("npc", "character", "party")) return;
+            sheet.form
+                .querySelectorAll(".inventory ul.items > li[data-item-id]")
+                .forEach((li) => {
+                    const id = (li as HTMLUListElement).dataset
+                        .itemId as string;
+                    const item = actor.items.get(id);
+                    if (!item || !item.isOfType("physical")) return;
+                    if (MonsterPart.hasMonsterPartData(item)) {
+                        const materials = new MonsterPart(item).materials;
+                        DynamicStyles.highlightElementOnHover(
+                            li,
+                            materials,
+                            "refined-item",
+                        );
+                    } else if (RefinedItem.hasRefinedItemData(item)) {
+                        const refinedItem = new RefinedItem(item);
+                        const materials = (
+                            AutomaticRefinementProgression.isEnabled
+                                ? refinedItem.imbuements
+                                : [
+                                      refinedItem.refinement,
+                                      ...refinedItem.imbuements,
+                                  ]
+                        ).map((m) => m.key);
+                        DynamicStyles.highlightElementOnHover(
+                            li,
+                            materials,
+                            "monster-part",
+                        );
+                    }
+                });
+        });
     }
 
     static updateItemsOnLevelChangeARP() {