+++ src/module.ts
@@ -4,7 +4,7 @@ import { MonsterPart } from "./monster-part";
 import { RefinedItem } from "./refined-item";
 import { Wrappers } from "./wrappers";
 import { ModuleHooks } from "./hooks";
-import { ActorPF2e } from "foundry-pf2e";
+import { ActorPF2e, ScenePF2e } from "foundry-pf2e";
 import { registerSF2eUuidRedirects } from "@src/uuid-redirect";
 import { API } from "@src/api";
 import { Compatibility } from "@src/compatibility";
@@ -27,28 +27,32 @@ Hooks.once("init", () => {
         Wrappers.extendItemRollOptions();
         ModuleHooks.registerAllHandlers();
 
+        // DynamicStyles.registerDragging();
+
         // refresh item data  for all actors
-        const refreshedItems: string[] = [];
-        [
-            ...game.actors.values(),
-            ...game.scenes.values().flatMap((sc: Scene) => {
-                return sc.tokens
-                    .values()
-                    .map((t) => t.actor as ActorPF2e | undefined);
-            }),
-        ].forEach((actor) => {
-            actor?.items.forEach((item) => {
-                if (
-                    item.isOfType("physical") &&
-                    (MonsterPart.hasMonsterPartData(item) ||
-                        RefinedItem.hasRefinedItemData(item)) &&
-                    !refreshedItems.includes(item.id)
-                ) {
-                    item.prepareDerivedData();
-                    refreshedItems.push(item.id);
-                }
+        if (game.user.isActiveGM) {
+            const refreshedItems: string[] = [];
+            [
+                ...game.actors.values(),
+                ...game.scenes.values().flatMap((sc: ScenePF2e) => {
+                    return sc.tokens
+                        .values()
+                        .map((t) => t.actor as ActorPF2e | undefined);
+                }),
+            ].forEach((actor) => {
+                actor?.items.forEach((item) => {
+                    if (
+                        item.isOfType("physical") &&
+                        (MonsterPart.hasMonsterPartData(item) ||
+                            RefinedItem.hasRefinedItemData(item)) &&
+                        !refreshedItems.includes(item.id)
+                    ) {
+                        item.prepareDerivedData();
+                        refreshedItems.push(item.id);
+                    }
+                });
             });
-        });
+        }
     });
     if (game.system.id === "sf2e") {
         Hooks.once("ready", registerSF2eUuidRedirects);